Add Star Wars iOS example client
