--- Examples/StarwarsExample/Package.resolved
@@ -1,15 +1,0 @@
-{
-  "originHash" : "affd180daa083d98d2852c57ee09e0f73057a971addf5d2796e88e4fc47438e5",
-  "pins" : [
-    {
-      "identity" : "swift-collections",
-      "kind" : "remoteSourceControl",
-      "location" : "https://github.com/apple/swift-collections",
-      "state" : {
-        "revision" : "a0cb0954ecb21e4e31b0070e6ed5674e8556685a",
-        "version" : "1.6.0"
-      }
-    }
-  ],
-  "version" : 3
-}
