--- Examples/StarwarsExample/Sources/StarwarsExample/SchemaTypes/Enums/Episode.graphqls.swift
@@ -1,8 +1,0 @@
-// @generated
-
-enum Episode: String, Encodable, Sendable {
-    case newHope = "NEW_HOPE"
-    case empire = "EMPIRE"
-    case jedi = "JEDI"
-}
-
