+++ StarwarsExample/client/Codegen/Package.resolved
@@ -1,0 +1,15 @@
+{
+  "originHash" : "2e1cfad5f5dcf868b900ac5706a915c0c6b63431df551191fde400e6ea4eb17b",
+  "pins" : [
+    {
+      "identity" : "swift-collections",
+      "kind" : "remoteSourceControl",
+      "location" : "https://github.com/apple/swift-collections",
+      "state" : {
+        "revision" : "a0cb0954ecb21e4e31b0070e6ed5674e8556685a",
+        "version" : "1.6.0"
+      }
+    }
+  ],
+  "version" : 3
+}
