+++ StarwarsExample/client/Packages/GraphQL/Operations/FavoriteEpisodeChangedSubscription.graphql
