+++ StarwarsExample/client/Packages/GraphQL/Operations/HeroQuery.graphql
