+++ StarwarsExample/client/Packages/GraphQL/Operations/SetFavoriteEpisodeMutation.graphql
