+++ StarwarsExample/client/Packages/GraphQL/Schema/Enums/Episode.graphqls.swift
@@ -1,0 +1,8 @@
+// @generated
+
+public enum Episode: String, Encodable, Sendable {
+    case newHope = "NEW_HOPE"
+    case empire = "EMPIRE"
+    case jedi = "JEDI"
+}
+
