+++ StarwarsExample/client/Packages/GraphQL/Schema/Scalars/ID.graphqls.swift
@@ -1,0 +1,5 @@
+// @generated
+// Any changes to this file will not be overwritten by future code generation execution.
+
+/// The `ID` scalar type represents a unique identifier, often used to refetch an object or as key for a cache. The ID type appears in a JSON response as a String; however, it is not intended to be human-readable. When expected as an input type, any string (such as `"4"`) or integer (such as `4`) input value will be accepted as an ID.
+public typealias ID = String
