+++ StarwarsExample/client/Packages/GraphQL/Utility/GraphQLEnum.swift
@@ -1,0 +1,15 @@
+// @generated
+
+public enum GraphQLEnum<T>: Decodable, Hashable, Sendable where T: Hashable & RawRepresentable & Sendable, T.RawValue == String {
+    case known(T)
+    case unknown(String)
+
+    public init(from decoder: Decoder) throws {
+        let rawValue = try decoder.singleValueContainer().decode(String.self)
+        if let value = T(rawValue: rawValue) {
+            self = .known(value)
+        } else {
+            self = .unknown(rawValue)
+        }
+    }
+}
