+++ StarwarsExample/client/Packages/GraphQL/Utility/HTTPSupport/GraphQLOperation.swift
@@ -1,0 +1,49 @@
+// @generated
+
+/// A `GraphQLOperation` represents a GraphQL document containing a single operation.
+public protocol GraphQLOperation: Sendable {
+
+    /// The optional name of the operation.
+    /// https://spec.graphql.org/September2025/#sel-FAFTDCFABAADFCBAAD-zM
+    static var operationName: String? { get }
+
+    /// The executable string operated on by a GraphQL service, containing
+    /// an operation definition and zero or more fragment definitions.
+    /// https://spec.graphql.org/September2025/#sec-Document
+    static var document: String { get }
+
+    /// The parameterized variables to execute the operation with.
+    /// https://spec.graphql.org/September2025/#sec-Language.Variables
+    var variables: Variables { get }
+
+    /// The operation's variables erased for request encoding.
+    /// Variable-free operations return `nil` so request encoders omit them.
+    var requestVariables: AnyEncodable? { get }
+
+    /// Metadata associated with the operation to include in the request.
+    var extensions: [String: AnyEncodable]? { get }
+
+    associatedtype Variables: Encodable, Sendable
+    associatedtype Data: Decodable, Sendable
+}
+
+extension GraphQLOperation {
+    public var requestVariables: AnyEncodable? {
+        let requestVariables: AnyEncodable = AnyEncodable(variables)
+        return requestVariables
+    }
+}
+
+extension GraphQLOperation where Variables == Never? {
+    public var requestVariables: AnyEncodable? { nil }
+}
+
+/// A `GraphQLSingleResponseOperation` produces one response and can be executed with `URLSession.request`.
+/// Queries and mutations have this capability; subscriptions produce a stream instead.
+public protocol GraphQLSingleResponseOperation: GraphQLOperation {}
+
+public protocol GraphQLQuery: GraphQLSingleResponseOperation {}
+
+public protocol GraphQLMutation: GraphQLSingleResponseOperation {}
+
+public protocol GraphQLSubscription: GraphQLOperation {}
