+++ StarwarsExample/client/iOS/StarwarsExampleApp.swift
@@ -1,0 +1,17 @@
+//
+//  StarwarsExampleApp.swift
+//  StarwarsExample
+//
+//  Created by Peter Meyers on 8/9/26.
+//
+
+import SwiftUI
+
+@main
+struct StarwarsExampleApp: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+        }
+    }
+}
